Add new "amttool" driver based on amttool.pl script
The "amtterm" project provides a simple Perl script for interacting
with AMT v7 and v8 hardware. This utility is available from the
"amtterm" package on Ubuntu 14.04 (and other distros), and was imported
previously in commit I6cfb4c804af00cfeb364017566e21b9c34cb6316

This commit implements simple Power and Management interfaces, and
bindings for PXE and Agent deploy methods.

Change-Id: I51fa7147a642c01953169edd61058d09e15a6e5f